pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "ranex-core"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "pydantic-settings",
    "sqlalchemy",
    "alembic",
    "typer",
    "rich",
    "PyYAML",
    "redis",
    "psycopg[binary]",
    "asyncpg",
    "mysql-connector-python",
    "PyMySQL",
    "passlib[bcrypt]",
    "python-multipart",
    "email-validator",
    "prometheus-client>=0.19.0",  # For metrics export
    "psutil>=5.9.0",  # For system resource monitoring (optional)
]

[project.scripts]
ranex = "ranex.cli:app"

[tool.maturin]
bindings = "pyo3"
python-packages = ["ranex"]
